Move windows to sub directory (WIP); Small fixes
